Changed Buttons
Ensured buttons worked to return to home page.
