--- app/src/main/java/com/example/taxilink/RideInformation/NearbyPlacesAdapter.java
@@ -1,50 +1,0 @@
-package com.example.taxilink.RideInformation;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-import java.util.List;
-import com.example.taxilink.R;
-
-public class NearbyPlacesAdapter extends RecyclerView.Adapter<NearbyPlacesAdapter.ViewHolder> {
-
-    private List<String> places;
-
-    public NearbyPlacesAdapter(List<String> places) {
-        this.places = places;
-    }
-
-    @NonNull
-    @Override
-    public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext())
-                .inflate(R.layout.nearby_place_item, parent, false);
-        return new ViewHolder(view);
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
-        String place = places.get(position);
-        holder.nameTextView.setText(place);
-    }
-
-    @Override
-    public int getItemCount() {
-        return places.size();
-    }
-
-    public static class ViewHolder extends RecyclerView.ViewHolder {
-        public TextView nameTextView;
-
-        public ViewHolder(@NonNull View itemView) {
-            super(itemView);
-            nameTextView = itemView.findViewById(R.id.place_name_textview);
-        }
-    }
-}
-
